Implement the back end of `CommandLine.executablePath`.
This PR implements the C++ back end of `CommandLine.executablePath` as approved
in [SE-0513](https://github.com/swiftlang/swift-evolution/blob/main/proposals/0513-commandline-executablepath.md).

The front end that exposes a `FilePath`-based API is blocked on the existence of
`FilePath` in the standard library, but some toolchain components like Swift
Testing need to back-deploy and can't link to the System package directly, so
they need to directly use the C++ implementation.