[Sanitizers] Add metadata to memsets created in zeroInit for LLDB
LLDB memsets new allocas to zero. It does this in order to be able to display a friendly 'variable is uninitialized' message rather than garbage. Unfortunately this use of a variable before its lifetime.start disagrees with the memtag-stack tagging pass.

This patch attaches a piece of metadata to these memsets, so that the memtag-stack tagging pass can recognize them and work around them appropriately.

rdar://162206592