[ConstraintSystem] Expand use of transitive conformance constraint
Produce a transitive conformance constraint to the result type of
a leading-dot syntax chain when it's converted to a type variable
that represents a generic parameter.

Record type variable that represents the base type of the chain
as referenced in the constraint, this makes it possible for the
constraint to be introduced as inference source for the type
variable.

This supports cases like: `let v: some P = .myValue` and
`func test<T: P>(_: T = .myValue) {}`