[cxx-interop] Suppress warning in non-arc Obj-C++
We have some arc specific casts in our generated headers that can
trigger some warnings in non-arc Obj-C++ code. These casts are noop in
that case so it is fine to have them there. And we do want them as we
want the code to be compatible both with arc and non-arc code.

rdar://163281971