Make LifetimeDependenceChecker::checkEnumElementDecl a static method
Lifetime dependence checking for enum elements is simple, and almost entirely
contained within this method. Making it a static method separates it from the
mostly unrelated code for AbstractFunctionDecl.

We could have made it a stand-alone function; this just minimises the diff size.

This is a preliminary step in refactoring LifetimeDependenceChecker to support
closures.