[cxx-interop] Support rvalue-ref qualified methods
We already imported methods that have && as qualifiers but did not
support them properly. When in the C++ code there is overloading based
on the reference qualifiers the imported Swift functions are ambiguous.
This patch append "Consuming" to the && qualified methods' names to
avoid the ambiguity. It also makes sure these methods are actually
imported as consuming.

In case of operators, however, Swift does not have a way to distinguish
between the consuming and non-consuming self. This patch prevents
importing the consuming variants. This fixes some compiler crashes
during deserialization that we hit due to the ambiguities.

Fixes #83801

rdar://158675235