[6.3][cxx-interop] Fix crash when virtual methods take move-only types
Explanation: We build forwarding methods to call the virtual methods. The forwarding methods tried to copy move-only types which resulted in a compiler crash. Now we try to detect this scenario and insert the required cast to make sure we get a move instead.
Issues: rdar://162195228
Original PRs: #85711
Risk: Low, we only add the cast to rvalue reference to types where we
could determine they are move only. So the changed code path should only
affect the problematic cases.
Testing: Added a compiler test.
Reviewers: @egorzhdan @susmonteiro @j-hui