[AST][ClangImporter] Add "equivalent" type, generate and use it.
When we import types from Clang, we may map some of them, for instance
`size_t` turns into `Int`, and `NSUInteger` will also turn into `Int`
in some circumstances.

This creates a problem because when we generate a Clang type _from_ a Swift
type, we will turn `Int` into `long`, not `size_t` or `NSUInteger`, so the Clang
type generated from the Swift type will not match the one we imported.

To fix this, we add to `ExtInfo` the notion of an "equivalent" type, which is
defined as the original type with all unsigned types changed to signed types.
We also at the same time replace top-level references with pointers, since we
won't care about the difference between those either in this context.

The equivalent type gets used to determine whether two function types are
incompatible at a Clang type level.