[AST] Elaborate ObjC protocol and class types.
In order to properly type check Clang function types, we need to know
the inheritance hierarchy and the set of protocols implemented by an
ObjC class.

Previously we were just creating empty `ObjCProtocolDecl` or
`ObjCInterfaceDecl`s, but that isn't sufficient and means that we won't be able
to properly support function pointer conversions.