[SILGen] Allow opaque reads to noncopyable properties to use getter
Currently OpaqueReadOwnershipRequest::evaluate always returns Borrowed for noncopyable types.
This results in calling code outside of the resilience domain being required to go through Read accessor
for accessing noncopyable properties, even if a getter was available.
This then prevents calling code from obtaining owned values from noncopyable acccessors, which is
required for making meaningful use of nonescapable types since they represent a projection of the source.

This can be fixed by having the resilient library specify OpaqueReadOwnership::Owned for read access to
noncopyable types if there was a getter specified. The result is that when compiling the calling code,
AbstractStorageDecl::requiresOpaqueReadCoroutine will now return false, allowing the calling code to use the getter
when querying  getOpaqueReadAccessStrategy)