[ConstraintGraph] Abstract adding constraints to inference for each node
There are two variants of `introduceToInference` now, one for a fixed
type binding and the other is for a new constraint added to the node.

This wasn't necessary before since it was a simple `infer` call on
`PotentialBindings` but the plan is to expand this into performing
the transitive inference through subtype chains as well.