Fix the BorrowingIteratorAdapter type
Because the BorrowingIteratorAdapter type still required Copyable
elements, it was failing to be allowed as the default BorrowIterator
type for any sequences. However, just marking the `Iterator.Element`
type as `~Copyable` is impossible, since IteratorProtocol has a
copyable element type, and switching to being generic over a sequence
with a noncopyable element type fails with this error:

Cannot suppress '~Copyable' on generic parameter 'S.Element' defined in outer scope

This change resolves things by adding `Element: ~Copyable` as a
second generic parameter and equating the new type and
`Iterator.Element`.