[CSBindings] Disambiguate Sendable and non-Sendable function type bindings
If the type variable prefers subtypes, diasambiguate a situation
when this type variable is simultaneously a supertype of `@Sendable`
function type and a subtype of a non-Sendable one by using a supertype
binding because it constitutes a "subtype" in this case.

For example:

```
@Sendable () -> Void conv $T
$T argument conv () -> Void
```

Either of the types could also be wrapped in a number of optionals. Even if
there is an optionality mismatch, let's still prefer a supertype binding
because that would be easier to diagnose.

In particular, this is helpful with ternary operators where the context is
non-Sendable, but one or both sides are.

Resolves: rdar://125948508
Resolves: rdar://136315240