Fix the swiftCxxStdlib link with explicit module build on Windows
We currently support static linking swiftCxxStdlib only on
Windows. When C++ interop is enabled but swiftCxxStdlib isn't actually
used or only when its underlying std module is used, we incorrectly
emitted the dynamic version of the lib name and caused the link
error. This change fixes it by always adding it to the dependency.

Issue https://github.com/swiftlang/swift/issues/85876