[cxx-interop] Use LookupQualifiedName in derived conformances
This is the first patch of a series of patches intended to shift lookup
toward using Clang lookup, to avoid Swift lookup's dependency on eager
Clang member importation.

The lookupCxxTypeMember() helper function replaces the previous helper
function, lookupNestedClangTypeDecl(). Functionally, the main difference
is that lookupCxxTypeMember() uses clang::Sema::LookupQualifiedName(),
which takes cares of issues like inheritance, ambiguity, etc. Meanwhile,
lookupNestedClangTypeDecl() only used clang::DeclContext::lookup(),
which is limited to the lexical context, and does not work with
inheritance.

The isIterator() function is renamed to hasIteratorConcept() to better
reflect what it is checking, and still uses clang::DeclContext::lookup()
to preserve the original behavior and to avoid requiring its callers to
supply a clang::Sema instance.