[cxx-interop] Import operators renamed via `swift_name` consistently
Normally Swift imports `T& operator*` as `var pointee: T`. However, when the C++ operator has `swift_name` attribute attached to it, the importing was inconsistent: the renamed function was marked as deprecated, and an additional undesired overload was created.

For instance:
```
const std::string& operator*() const SWIFT_NAME(dereference());
```
was imported as an unavailable/deprecated `func dereference()`, and an additional overload of `func *(_:)` was created.

rdar://163500978