[6.3][cxx-interop] Fix crashing on recursive enums
Explanation: Checking whether an Enum can be exported to C++ included
checking whether all the enum cases are exportable. When one of these
cases refer to the original enum we got into an infinte recursion. This
PR adds a cache that is also used to cut the recursion in these cases.
Issues: rdar://164153038
Original PRs: #85980
Risk: Low, this only adds a cache.
Testing: Added a compiler test.
Reviewers: @egorzhdan @j-hui @susmonteiro